refactor: update ThemeProvider import paths and add ThemeProvider component
