+++ src/App.tsx
@@ -2,7 +2,7 @@ import { Provider } from "react-redux";
 import { persistor, store } from "@/store/store";
 import MainLayout from "./layout/MainLayout";
 import WeatherDashboard from "./components/Home";
-import { ThemeProvider } from "./components/theme-provider";
+import { ThemeProvider } from "./providers/ThemeProvider";
 import { PersistGate } from "redux-persist/integration/react";
 
 function App() {
